+++ ui_automation_tests/pages/flags_list_page.py
@@ -4,7 +4,6 @@ from shared.BasePage import BasePage
 
 class FlagsListPage(BasePage):
     BUTTON_ADD_FLAG_ID = "button-add-a-flag"
-    INPUT_NAME_FILTER_ID = "name"
     CHECKBOX_ONLY_SHOW_DEACTIVATED_NAME = "status"
 
     def click_add_a_flag_button(self):
@@ -12,12 +11,6 @@ class FlagsListPage(BasePage):
 
     def click_edit_link(self):
         self.driver.find_element_by_partial_link_text("Edit").click()
-
-    def filter_by_name(self, name):
-        functions.try_open_filters(self.driver)
-        self.driver.find_element_by_id(self.INPUT_NAME_FILTER_ID).clear()
-        self.driver.find_element_by_id(self.INPUT_NAME_FILTER_ID).send_keys(name)
-        self.driver.find_element_by_css_selector(selectors.BUTTON_APPLY_FILTERS).click()
 
     def click_only_show_deactivated(self):
         functions.try_open_filters(self.driver)
